milestone 2/requirements.txt
numpy
pandas
scikit-learn
requests
beautifulsoup4
fastparquet
kafka-python
psycopg2-binary
SQLAlchemy
docker

